Add global keymap
This is the first step in a fully-functional keymap system like that
in Emacs.  When complete, it will be possible to bind hotkeys to key
sequences.  Support for Emacs-style numeric prefix arguments is also
planned.

At present, the global keymap is in place, but there is not yet
support for contextual keymaps (e.g., application-specified keymaps).
Until such support exists, hotkeys using the global keymap will not be
compatible with existing contextual hotkeys, so keymap support remains
disabled.