Add `exit` keymap to RSSOwl configuration
My initial configuration attempted to use RSSOwl's internal key
bindings as much as possible; since it supports prefix keys, it seemed
like `C-x C-c` for exit would be straightforward.  However, the use of
`C-x` as a prefix key in the `navigation` keymap (which is
indispensible, since the up and down functions don't seem to be
bindable, and mark control would need to be handled in any case)
interferes with this by absorbing the initial `C-x` keystroke.

In the future, I might consider adjusting the prefix key logic so that
an unrecognized key sequence would emit its prefix keys also, but I'm
not convinced that wouldn't have unexpected effects, and simply adding
the `exit` keymap solves the immediate problem.