Fix MediaMonkey hotkeys properly
After further investigation, I realized that the key problem with the
editing hotkeys not working was just that the properties dialog had
changed its window name.  The new name is pretty generic ("chromium"),
so I also added an `ahk_exe` condition to ensure it remains specific to
MediaMonkey.

There also seemed to be a problem getting MediaMonkey to respond
properly to Shift-Enter (to open the properties dialog), but only when
using the right shift key.  I was unable to find an easy workaround and
I don't believe I'm using that hotkey anyway, so we'll just blacklist it
for all keymaps.